rename again!
generic renderer API is hard
