+++ html/html_test.go
@@ -1,0 +1,1 @@
+package html_test
