+++ usfm.go
@@ -5,13 +5,13 @@ import (
 	"os"
 	"strings"
 
-	"github.com/baijum/usfm/renderer"
+	"github.com/baijum/usfm/html"
 )
 
 func main() {
-	o := renderer.Options{}
+	o := html.Options{}
 	s := ""
-	html := renderer.NewHTML(o, strings.NewReader(s))
+	html := html.New(o, strings.NewReader(s))
 	f, _ := os.Create("out.html")
 	defer f.Close()
 	err := html.Render(f)
